test(worker): include streaming negotiation in existing CI history suite
Copilot-Session: e0af01a5-0dfa-4e71-a660-c4186e65d7e0
